rename rust_deps to rust_syncer
